% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_2010.R
\docType{data}
\name{census_2010}
\alias{census_2010}
\title{Census 2010}
\format{
\subsection{census_2010}{

A data frame with 73057 obs. and  3 variables
\describe{
\item{GEOID}{Geographical Identifiers}
\item{NAME}{name of census tract county state}
\item{pop10}{population}
}
}
}
\source{
\url{https://github.com/ruralinnovation/senior-data-engineer-assessment/blob/main/data/census_tract_populations_2010.csv}
}
\usage{
census_2010
}
\description{
A subset of US 2010 census tract with only population count
}
\keyword{datasets}
