% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_2020.R
\docType{data}
\name{census_2020}
\alias{census_2020}
\title{Census 2020}
\format{
\subsection{census_2020}{

A data frame with 84414 obs. and  3 variables
\describe{
\item{GEOID}{Geographical Identifiers}
\item{NAME}{name of census tract county state}
\item{pop20}{population}
}
}
}
\source{
\url{https://github.com/ruralinnovation/senior-data-engineer-assessment/blob/main/data/census_tract_populations_2020.csv}
}
\usage{
census_2020
}
\description{
A subset of US 2020 census tract with only population count
}
\keyword{datasets}
