% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{land_change}
\alias{land_change}
\title{Calculate surface difference between 2020 tract and 2010}
\usage{
land_change(x, perc = TRUE)
}
\arguments{
\item{x}{is a tab20_tract20_tract10}

\item{perc}{if not specified will be TRUE to generate percentage if FALSE will generate raw units}
}
\description{
Use tab20_tract20-tract10 to calculate difference in  total land area between 2020 and 2010 and return a vector
}
