% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract20_10.R
\docType{data}
\name{tract20_10}
\alias{tract20_10}
\title{2020 Census Tract to 2010 Census Tract Relationship File}
\format{
\subsection{tract20_10}{

A data frame with 126450 obs. and  8 variables
\describe{
\item{GEOID_TRACT_20}{Geographical Identifiers for census tract 2020}
\item{AREALAND_TRACT_20}{Surface of land in 2020 (unit?)}
\item{AREAWATER_TRACT_20}{Surface of water in 2020 (unit?)}
\item{GEOID_TRACT_10}{Geographical Identifiers for census tract 2010}
\item{AREALAND_TRACT_10}{Surface of land in 2010 (unit?)}
\item{AREAWATER_TRACT_10}{Surface of water in 2010 (unit?)}
\item{AREALAND_PART}{guessing}
\item{AREAWATER_PART}{guessing}
}
}
}
\source{
\url{https://www2.census.gov/geo/pdfs/maps-data/data/rel2020/tract/explanation_tab20_tract20_tract10.pdf}
}
\usage{
tract20_10
}
\description{
A subset of 2020 Census Tract to 2010 Census Tract Relationship File
}
\keyword{datasets}
